agent/requirements.txt
httpx
ollama
python-dotenv
python-telegram-bot
thefuzz
python-Levenshtein
